+++ sphae.sh
@@ -11,7 +11,7 @@
 ##SBATCH --partition=high-capacity
 ##SBATCH --qos=hc-concurrent-jobs
 
-#sphae install
+sphae install
 export VVDB=sphae/workflow/databases/Pfam35.0/Pfam-A.hmm.gz
 export CHECKVDB=sphae/workflow/databases/checkv-db-v1.5
 export PHAROKKADB=sphae/workflow/databases/pharokka_db
@@ -20,5 +20,5 @@ export PHOLDDB=sphae/workflow/databases/phold
 
 #sphae run --input tests/data/illumina-subset --threads 32 -k --use-conda --conda-frontend mamba 
 #sphae run --input tests/data/nanopore-subset --sequencing longread --threads 64 -k --use-conda --conda-frontend mamba 
-sphae run --input tests/data/nanopore-subset --sequencing longread --threads 64 -k --no_medaka --use-conda --conda-frontend mamba 
-sphae annotate --genome tests/data/genome --threads 64 -k --use-conda --conda-frontend mamba
+#sphae run --input tests/data/nanopore-subset --sequencing longread --threads 64 -k --no_medaka --use-conda --conda-frontend mamba 
+#sphae annotate --genome tests/data/genome --threads 64 -k --use-conda --conda-frontend mamba