+++ Changes.md
@@ -1,7 +1,7 @@
 # Changes being added
 ## v1.4.7 
 - checkv database error fixed
-- updating phold.yaml file 
+- updating phold.yaml and medaka.yaml file 
 - updated to set the databases to variables that need to be set by the user
   - this to so the users download only those databases they dont have
   - `sphae install` still works to install all the required databases, but the variables still have to be set to run `sphae run`